updated Etiam

+++ .Projects/build.bat
@@ -1,18 +1,14 @@
 @echo off
-if exist "..\EXEs\%~n1" (
-  rd /s /q "..\EXEs\%~n1"
-)
-if exist "..\ZIPs\%~n1.zip" (
-  del "..\ZIPs\%~n1.zip"
-)
+if [%1] == [] goto nope
+if exist "..\EXEs\%~n1" rd /s /q "..\EXEs\%~n1"
+if exist "..\ZIPs\%~n1.zip" del "..\ZIPs\%~n1.zip"
+if exist "..\LOVEs\%~n1.love" del "..\LOVEs\%~n1.love"
 md "..\EXEs\%~n1"
 echo **ZIPPING FOLDER INTO .LOVE FILE**
 7z a "%~n1.zip" ".\%~n1\*" -r
-move "%~n1.zip" "%~n1.love"
+move "%~n1.zip" "..\LOVEs\%~n1.love"
 echo **CREATING .EXE FILE**
-taskkill /F /IM 7z.exe /T
-copy /b "love.exe"+"%~n1.love" "..\EXEs\%~n1\%~n1.exe"
-del "%~n1.love"
+copy /b "love.exe"+"..\LOVEs\%~n1.love" "..\EXEs\%~n1\%~n1.exe"
 if exist "%~n1\versioninfo.res" (
   echo **ADDING VERSIONINFO TO .EXE FILE**
   resourcehacker -addoverwrite "..\EXEs\%~n1\%~n1.exe", "..\EXEs\%~n1\%~n1.exe", "%~n1\versioninfo.res",,,
@@ -35,3 +31,8 @@ del "DLLs.zip"
 cd "..\..\.Projects"
 echo **ZIPPING INTO ..\ZIPs**
 7z a "..\ZIPs\%~n1.zip" "..\EXEs\%~n1" -r
+goto end
+:nope
+echo **NO FOLDER PROVIDED**
+pause
+:end
